mbc1, mbc2 y mbc3 funcionamient básico implementado

+++ src/gb/utils/BitOperations.java
@@ -2,15 +2,15 @@ package gb.utils;
 
 public class BitOperations {
 	
-	public static boolean testBit(byte b, int bit) {
+	public static boolean testBit(int b, int bit) {
 		return ((b & (1 << bit )) == (1 << bit));
 	}
 	
-	public static byte bitReset(byte b, int bit) {
+	public static byte bitReset(int b, int bit) {
 		return (byte) (b & ~(1 << bit));
 	}
 	
-	public static byte bitSet(byte b, int bit) {
+	public static byte bitSet(int b, int bit) {
 		return (byte) (b | (1 << bit));
 	}
